+++ src/Console/ApplicationFactory.php
@@ -4,19 +4,27 @@ declare(strict_types=1);
 
 namespace Mgrunder\Fuzz\Console;
 
+use Mgrunder\Fuzz\Console\Command\KillClientsCommand;
 use Mgrunder\Fuzz\Console\Command\WorkCommand;
 use Mgrunder\Fuzz\Fuzz\Command\CommandRegistry;
+use Mgrunder\Fuzz\Runtime\ClientKiller;
 use Mgrunder\Fuzz\Runtime\RelayClientFactory;
+use Mgrunder\Fuzz\Runtime\RelayAdminClientFactory;
 use Mgrunder\Fuzz\Runtime\RelayStatsProvider;
 use Mgrunder\Fuzz\Runtime\WorkerOrchestrator;
 use Symfony\Component\Console\Application;
 
 final class ApplicationFactory
 {
-    public function create(): Application
+    public function create(string $defaultCommand = WorkCommand::NAME): Application
     {
         $application = new Application('fuzz');
         $application->addCommands([
+            new KillClientsCommand(
+                new ClientKiller(
+                    new RelayAdminClientFactory(),
+                ),
+            ),
             new WorkCommand(
                 new WorkerOrchestrator(
                     new RelayClientFactory(),
@@ -25,7 +33,7 @@ final class ApplicationFactory
                 ),
             ),
         ]);
-        $application->setDefaultCommand(WorkCommand::NAME, true);
+        $application->setDefaultCommand($defaultCommand, true);
 
         return $application;
     }